requirements.txt
# primary
numpy
matplotlib
opencv
tqdm


# advanced(optional)
torch
torchvision
umap-learn
scikit-learn